requirements.txt
slither-analyzer>=0.10.0
networkx>=3.0
numpy>=1.24.0
solc-select>=1.0.0